Improve website responsiveness and layout design
The commit includes the redesigning of the SpecialCard, header, and hero components for enhancing UI and improving responsiveness on medium-sized screens. It also optimizes the site's CSS by modifying properties like margin, padding, width, height, and color for better visual appeal. Lastly, changes were made to navigate desktop JSX, and a loop was added to map over an array of specials data.